refactor: reorganiza estrutura de pacotes e melhora qualidade do código
- Renomeia pacotes para seguir convenções de clean architecture (adapters/in, adapters/out)
- Atualiza Spring Boot para versão 3.4.6
- Altera porta do MongoDB para 27019 no docker-compose
- Refatora casos de uso para implementar interfaces de ports
- Melhora tratamento de Optional nos repositórios
- Padroniza respostas dos controllers com DTOs
- Adiciona tratamento adequado de logs e erros
- Limpa classes de teste e atualiza referências de pacotes
