+++ docker-compose.yml
@@ -38,7 +38,7 @@ services:
       MONGO_INITDB_ROOT_PASSWORD: MongoDB2025!
       MONGO_INITDB_DATABASE: customer-service
     ports:
-      - "27017:27017"
+      - "27019:27017"
     volumes:
       - mongo-volume:/data/db
       - ./mongo-init.js:/docker-entrypoint-initdb.d/mongo-init.js:ro
