+++ pom.xml
@@ -6,7 +6,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>3.4.3</version>
+		<version>3.4.6</version>
 		<relativePath />
 	</parent>
 	<groupId>br.matosit</groupId>
