+++ src/main/java/br/matosit/customer_service/adapters/in/rest/responses/CustomerResponse.java
@@ -1,0 +1,7 @@
+package br.matosit.customer_service.adapters.in.rest.responses;
+
+import java.util.List;
+
+public record CustomerResponse(String id, String name, String email, String phone,
+    List<AddressResponse> addresses) {
+}
