--- src/main/java/br/matosit/customer_service/application/ports/CustomerRepository.java
@@ -1,15 +1,0 @@
-package br.matosit.customer_service.application.ports;
-
-import java.util.Optional;
-import br.matosit.customer_service.domain.entities.Customer;
-
-public interface CustomerRepository {
-
-  Customer save(Customer customer);
-
-  Optional<Customer> findById(String id);
-
-  void deleteById(String id);
-
-  Customer findByEmail(String email);
-}
