+++ src/main/java/br/matosit/customer_service/application/ports/in/CreateCustomerUseCase.java
@@ -1,0 +1,7 @@
+package br.matosit.customer_service.application.ports.in;
+
+import br.matosit.customer_service.domain.entities.Customer;
+
+public interface CreateCustomerUseCase {
+  Customer create(Customer customer);
+}
