+++ src/main/java/br/matosit/customer_service/application/ports/in/DeleteCustomerUseCase.java
@@ -1,0 +1,7 @@
+package br.matosit.customer_service.application.ports.in;
+
+public interface DeleteCustomerUseCase {
+
+  void delete(String id);
+
+}
