+++ src/main/java/br/matosit/customer_service/application/ports/in/FindCustomerUseCase.java
@@ -1,0 +1,9 @@
+package br.matosit.customer_service.application.ports.in;
+
+import br.matosit.customer_service.domain.entities.Customer;
+
+public interface FindCustomerUseCase {
+
+  Customer find(String id);
+
+}
