--- src/main/java/br/matosit/customer_service/presentation/handlers/GlobalExceptionHandler.java
@@ -1,57 +1,0 @@
-package br.matosit.customer_service.presentation.handlers;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import br.matosit.customer_service.domain.exceptions.CustomerAlreadyExistsException;
-import br.matosit.customer_service.domain.exceptions.CustomerNotFoundException;
-import br.matosit.customer_service.domain.exceptions.DomainException;
-import br.matosit.customer_service.presentation.responses.ErrorResponse;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-  @ExceptionHandler(CustomerAlreadyExistsException.class)
-  public ResponseEntity<ErrorResponse> handleCustomerAlreadyExists(
-      CustomerAlreadyExistsException ex) {
-    ErrorResponse response = ErrorResponse.of(ex.getCode(), ex.getMessage());
-    return ResponseEntity.status(HttpStatus.CONFLICT).body(response);
-  }
-
-  @ExceptionHandler(CustomerNotFoundException.class)
-  public ResponseEntity<ErrorResponse> handleCustomerNotFound(CustomerNotFoundException ex) {
-    ErrorResponse response = ErrorResponse.of(ex.getCode(), ex.getMessage());
-    return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
-  }
-
-
-  @ExceptionHandler(DomainException.class)
-  public ResponseEntity<ErrorResponse> handleDomainException(DomainException ex) {
-    ErrorResponse response = ErrorResponse.of(ex.getCode(), ex.getMessage());
-    return ResponseEntity.badRequest().body(response);
-  }
-
-  @ExceptionHandler(MethodArgumentNotValidException.class)
-  public ResponseEntity<ErrorResponse> handleValidationExceptions(
-      MethodArgumentNotValidException ex) {
-    ErrorResponse response = ErrorResponse.of("VALIDATION-001", "Erro de validação nos campos");
-
-    ex.getBindingResult().getAllErrors().forEach(error -> {
-      String fieldName = ((FieldError) error).getField();
-      String errorMessage = error.getDefaultMessage();
-      response.addValidationError(fieldName, errorMessage);
-    });
-
-    return ResponseEntity.badRequest().body(response);
-  }
-
-  @ExceptionHandler(Exception.class)
-  public ResponseEntity<ErrorResponse> handleGenericException(Exception ex) {
-    ErrorResponse response =
-        ErrorResponse.of("INTERNAL-001", "Ocorreu um erro interno no servidor");
-    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
-  }
-}
